restructed type files so they are closer to source code
